+++ inst/templates/ad_adtte.R
@@ -28,15 +28,15 @@ adsl <- adsl %>%
   derive_vars_merged(
     dataset_add = adrs,
     filter_add = PARAMCD == "RSP" & AVALC == "Y" & ANL01FL == "Y",
-    by_vars = vars(STUDYID, USUBJID),
-    new_vars = vars(TEMP_RESPDT = ADT)
+    by_vars = exprs(STUDYID, USUBJID),
+    new_vars = exprs(TEMP_RESPDT = ADT)
   )
 
 adrs <- adrs %>%
   derive_vars_merged(
     dataset_add = adsl,
-    by_vars = vars(STUDYID, USUBJID),
-    new_vars = vars(TEMP_RESPDT)
+    by_vars = exprs(STUDYID, USUBJID),
+    new_vars = exprs(TEMP_RESPDT)
   )
 
 # Use pre-defined tte_source objects to derive Overall Survival, Progression
@@ -47,7 +47,7 @@ adtte <- derive_param_tte(
   event_conditions = list(death_event),
   censor_conditions = list(lastalive_censor, rand_censor),
   source_datasets = list(adsl = adsl, adrs = adrs),
-  set_values_to = vars(PARAMCD = "OS", PARAM = "Overall Survival")
+  set_values_to = exprs(PARAMCD = "OS", PARAM = "Overall Survival")
 ) %>%
   derive_param_tte(
     dataset_adsl = adsl,
@@ -55,7 +55,7 @@ adtte <- derive_param_tte(
     event_conditions = list(pd_event, death_event),
     censor_conditions = list(lasta_censor, rand_censor),
     source_datasets = list(adsl = adsl, adrs = adrs),
-    set_values_to = vars(PARAMCD = "PFS", PARAM = "Progression Free Survival")
+    set_values_to = exprs(PARAMCD = "PFS", PARAM = "Progression Free Survival")
   ) %>%
   derive_param_tte(
     dataset_adsl = adsl,
@@ -66,7 +66,7 @@ adtte <- derive_param_tte(
       adsl = filter(adsl, !is.na(TEMP_RESPDT)),
       adrs = filter(adrs, !is.na(TEMP_RESPDT))
     ),
-    set_values_to = vars(PARAMCD = "RSD", PARAM = "Duration of Response")
+    set_values_to = exprs(PARAMCD = "RSD", PARAM = "Duration of Response")
   )
 
 # Derive analysis value and sequence
@@ -77,8 +77,8 @@ adtte <- adtte %>%
     end_date = ADT
   ) %>%
   derive_var_obs_number(
-    by_vars = vars(STUDYID, USUBJID),
-    order = vars(PARAMCD),
+    by_vars = exprs(STUDYID, USUBJID),
+    order = exprs(PARAMCD),
     check_type = "error"
   )
 
@@ -86,8 +86,8 @@ adtte <- adtte %>%
 adtte <- adtte %>%
   derive_vars_merged(
     dataset_add = adsl,
-    new_vars = vars(ARMCD, ARM, ACTARMCD, ACTARM, AGE, SEX),
-    by_vars = vars(STUDYID, USUBJID)
+    new_vars = exprs(ARMCD, ARM, ACTARMCD, ACTARM, AGE, SEX),
+    by_vars = exprs(STUDYID, USUBJID)
   )
 
 # ---- Save output ----