+++ NEWS.md
@@ -1,3 +1,12 @@
+# admiralonco 0.3.0
+
+## Breaking Changes
+
+- All function arguments which expected a list of quosures created by `vars()`
+are now expecting a list of expressions created by `exprs()`. For example,
+instead of `by_vars = vars(STUDYID, USUBJID)` `by_vars = exprs(STUDYID,
+USUBJID)` must be used now.
+
 # admiralonco 0.2.0
 
 ## Updates of Existing Functions